Removed research files, moved luts/netcdf.py to utils/luts.py, updated imports to match
